feat: add validation on form custumer

+++ src/pages/checkout/Checkout.page.jsx
@@ -18,6 +18,23 @@ import Modal from "../../components/modal/Modal.component";
 function CheckoutPage() {
   const navigate = useNavigate();
   const [isShowModal, setIsShowModal] = useState(false);
+  const [isShowModalError, setIsShowModalError] = useState(false);
+  const [formData, setFormData] = useState({
+    firstName: "",
+    lastName: "",
+    address: "",
+    cp: "",
+    email: "",
+    phone: "",
+    notes: "",
+  });
+
+  const handleChange = ({ fieldName, event }) => {
+    setFormData({
+      ...formData,
+      [fieldName]: event.currentTarget.value,
+    });
+  };
 
   const handleNavigate = (path) => {
     navigate(`${path}`);
@@ -27,19 +44,42 @@ function CheckoutPage() {
     setIsShowModal(false);
   };
 
+  const handleCloseModalError = () => {
+    setIsShowModalError(false);
+  };
+
   const handleOrder = () => {
-    setIsShowModal(true);
-    window.localStorage.clear();
-    setTimeout(function () {
-      setIsShowModal(false);
-      navigate("/home");
-    }, 2000);
+    console.log(formData);
+    if (
+      formData.firstName.length > 0 &&
+      formData.lastName.length > 0 &&
+      formData.cp.length > 0 &&
+      formData.phone.length > 0 &&
+      formData.email.length > 0
+    ) {
+      setIsShowModal(true);
+      window.localStorage.clear();
+      setTimeout(function () {
+        setIsShowModal(false);
+        navigate("/home");
+      }, 2000);
+    } else {
+      setIsShowModalError(true);
+      setTimeout(function () {
+        setIsShowModal(false);
+      }, 2000);
+    }
   };
 
   return (
     <>
       <Modal show={isShowModal} handleClose={handleCloseModal}>
         <StyledLabelAdded>Order created</StyledLabelAdded>
+      </Modal>
+      <Modal show={isShowModalError} handleClose={handleCloseModalError}>
+        <StyledLabelAdded>
+          Please complete the information required.
+        </StyledLabelAdded>
       </Modal>
       <StyledWrapperReturns>
         <StyledReturns onClick={() => handleNavigate("/home")}>
@@ -52,7 +92,10 @@ function CheckoutPage() {
       </StyledWrapperReturns>
       <StyledWrapper>
         <StyledWrapperLeft>
-          <FormCustumerInformation />
+          <FormCustumerInformation
+            handleChange={handleChange}
+            formData={formData}
+          />
         </StyledWrapperLeft>
         <StyledWrapperRight>
           <TableCheckout />
